Move priority picker to shared controls
